fix(worklist): reach survives the rebase, and the fixtures carry it
The reach tests moved into their own file — worklist_test.go crossed the
1000-line test ceiling.

Two frontend maps gained capture_confidentiality_verdict, which arrived on
main while this branch was open and which every AI task must appear in.

The three story fixtures and the test day helper carry reach, which the
contract now requires.
