+++ frontend/src/screens/worklist.test.tsx
@@ -63,6 +63,7 @@ function day(over: Partial<Worklist> = {}): Worklist {
     queue: [],
     summary: { urgent: 0, due: 0, lower_priority: 0, total: 0 },
     sources_unavailable: [],
+    reach: [],
     ...over,
   };
 }
